% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_base_precise.R
\name{bs_bbase_precise}
\alias{bs_bbase_precise}
\title{Get B-splines. Written by Dr. Niamh Cahill.}
\usage{
bs_bbase_precise(
  x = x,
  lastobs = max(x),
  xl = min(x),
  xr = max(x),
  nseg = 10,
  deg = 3
)
}
\arguments{
\item{x}{Vector of x-values (without NAs) for which splines need to be calculated (determines the number of rows of B.ik)}

\item{lastobs}{= max(x) max x-value which determines knot placement. Aligns knot to most recent survey year.}

\item{xl}{= min(x) min x-value which determines knot placement.}

\item{xr}{= max(x) max x-value which determines knot placement}

\item{nseg}{Number of segments to divide x vector into, used to place knots}

\item{degree}{= 3 # currently tested only with degree 3}
}
\value{
returns B.ik matrix, each row is one observation, each column is one B-spline. Vector of knots indicating knot placement across x vector.
}
\description{
Get B-splines. Written by Dr. Niamh Cahill.
}
