% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_sigma_delta_hat_logitnormal.R
\docType{data}
\name{Bspline_sigma_matrix_median}
\alias{Bspline_sigma_matrix_median}
\title{The median estimate for the national-level variance-covariance matrix of the delta.k terms in the multi-country national model.
This array is used to inform the Wishart prior of the single-country national model.}
\format{
A array of 2 matrices with 5 rows and 5 columns:
\describe{
\item{Array 1}{Estimated public sector variance-covaraince matrix}
\item{Array 2}{Estimated private sector variance-covaraince matrix}
}
}
\usage{
Bspline_sigma_matrix_median
}
\description{
The median estimate for the national-level variance-covariance matrix of the delta.k terms in the multi-country national model.
This array is used to inform the Wishart prior of the single-country national model.
}
\keyword{Bspline_sigma_matrix_median}
