% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format.R
\name{check_format}
\alias{check_format}
\title{Check format of data}
\source{
With thanks, taken from https://github.com/AlkemaLab/fpemlocal/blob/master/R/format_check.R
}
\usage{
check_format(format_list, data)
}
\arguments{
\item{format_list}{The list of requirements for the data type (national or subnational). See the /data folder to review the _format files.}

\item{data}{The data to be checked.}
}
\value{
An informative error message if data does not pass validation
}
\description{
Check format of data
}
