% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_subnational_names.R
\name{clean_subnat_names}
\alias{clean_subnat_names}
\title{Get JAGS model inputs}
\usage{
clean_subnat_names(fp2030 = TRUE, raw_subnatdata)
}
\arguments{
\item{fp2030}{TRUE/FALSE. Default is TRUE. Filters the data to only include FP2030 countries.}

\item{raw_subnatdata}{The subnational family planning source data from the 'get_subnational_data' function.}
}
\value{
A dataset with the subnational regions names cleaned for Rwanda, Nigeria and Cote d'Ivoire. Optional to filter for only FP2030 countries.
}
\description{
Get JAGS model inputs
}
\examples{
Include only FP2030 countries: subnat_data <- clean_subnat_names(fp2030=TRUE, raw_subnatdata)
Include all countries: subnat_data <- clean_subnat_names(fp2030=FALSE, raw_subnatdata). Note this functionality is untested.
}
