% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Country_and_area_classification.R
\docType{data}
\name{Country_and_area_classification}
\alias{Country_and_area_classification}
\title{Track20 Country and area classification according to the United Nations
A subset of data from the United Nations country classifications}
\format{
A data frame with 231 rows and 8 columns:
\describe{
\item{Country or area}{Country name}
\item{ISO Code}{1, 2 & 3 number ISO country codes}
\item{Major area}{Continent}
\item{Region}{Sub-continent}
\item{Developed region}{Binary indicator for development status}
\item{Least developed country}{Binary indicator for least developed status}
\item{Sub-Saharan Africa}{Binary indicator for whether country is in Sub-Saharan Africa}
\item{FP2020}{Binary indicator for FP2020 participation status}
}
}
\source{
On request from Track20. Contact details found at https://www.track20.org/
}
\usage{
Country_and_area_classification
}
\description{
Track20 Country and area classification according to the United Nations
A subset of data from the United Nations country classifications
}
\keyword{Country_and_area_classification}
