% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Country_and_area_classification_inclFP2020.R
\docType{data}
\name{Country_and_area_classification_inclFP2020}
\alias{Country_and_area_classification_inclFP2020}
\title{Track20 Country and area classification according to the United Nations}
\format{
## `Country_and_area_classification_inclFP2020`
A data frame with 231 rows and 8 columns:
\describe{
  \item{Country or area}{Country name}
  \item{ISO Code}{1, 2 & 3 number ISO country codes}
  \item{Major area}{Continent}
  \item{Region}{Sub-continent}
  \item{Developed region}{Binary indicator for development status}
  \item{Least developed country}{Binary indicator for least developed status}
  \item{Sub-Saharan Africa}{Binary indicator for whether country is in Sub-Saharan Africa}
  \item{FP2020}{Binary indicator for FP2020 participation status}
}
}
\source{
On request from Track20. Contact details found at https://www.track20.org/
}
\usage{
Country_and_area_classification_inclFP2020
}
\description{
A subset of data from the United Nations country classifications
}
\keyword{datasets}
