% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_index.R, R/country_index_fun.R
\name{country_index_fun}
\alias{country_index_fun}
\alias{continent_index_fun}
\title{Indexing function for countries used in data.}
\usage{
country_index_fun(my_data, my_country)

continent_index_fun(my_data)

country_index_fun(my_data, my_country)
}
\arguments{
\item{my_data}{The sub-national family planning source data for the country of interest.}

\item{my_country}{The name of your country of interest.}

\item{my_countries}{A vector of country names used in the data}
}
\value{
returns table with countries indexed

Dataframe with country indexing applied
}
\description{
Indexing function for countries used in data.

Apply country indexing to data
}
\section{Functions}{
\itemize{
\item \code{country_index_fun()}: Calculates the country index for named countries

\item \code{continent_index_fun()}: Calculates the continent index for named continents

}}
\examples{
country_index_fun(mydata, "Nepal")
}
