% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_names.R
\docType{data}
\name{country_names}
\alias{country_names}
\title{The names of the countries with national and subnational administration level data stored}
\format{
A data frame with 30 rows and 3 columns:
\describe{
\item{Country names}{Country name}
\item{National level data available}{Indicator for whether data at the national administration level is available in the package}
\item{Subnational level data available}{Indicator for whether data at the subnational administration level is available in the package}
}
}
\usage{
country_names
}
\description{
The names of the countries with national and subnational administration level data stored
}
\keyword{country_names}
