% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEFT_DHS_database.R
\docType{data}
\name{DEFT_DHS_database}
\alias{DEFT_DHS_database}
\title{DEFT_DHS_database
A database of the design effects for some of the DHS surveys in the national and subnational datasets.
Due to due to multistage and clustering of the DHS sample, the average standard error is increased by a design effect (DEFT) factor over that in an equivalent simple random sample.}
\format{
A dataframe of Country names, survey year and design effects for DHS surveys
}
\source{
DHS final reports Appendix B, 'ESTIMATES OF SAMPLING ERRORS'.
}
\usage{
DEFT_DHS_database
}
\description{
DEFT_DHS_database
A database of the design effects for some of the DHS surveys in the national and subnational datasets.
Due to due to multistage and clustering of the DHS sample, the average standard error is increased by a design effect (DEFT) factor over that in an equivalent simple random sample.
}
\keyword{DEFT_DHS_database}
