% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimated_global_subnational_correlations.R
\docType{data}
\name{estimated_global_subnational_correlations}
\alias{estimated_global_subnational_correlations}
\title{The estimated subnational-level correlations between the rates of change in methods}
\format{
A array of 2 matrices with 5 rows and 5 columns:
\describe{
\item{row}{Contraceptive method the correlation estimate refers to}
\item{column}{Contraceptive method the correlation estimate refers to}
\item{public_cor}{The estimated correlation between the rates of change in methods supplied by the public sector}
\item{private_cor}{The estimated correlation between the rates of change in methods supplied by the private sector}
}
}
\usage{
estimated_global_subnational_correlations
}
\description{
The estimated subnational-level correlations between the rates of change in methods
}
\keyword{estimated_global_subnational_correlations}
