% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_cor_matrix.R
\name{flat_cor_mat}
\alias{flat_cor_mat}
\title{Flatten correlation matrix}
\usage{
flat_cor_mat(cor_r)
}
\arguments{
\item{cor_r}{Correlation matrix}
}
\value{
A table with 3 columns containing :
}
\description{
Flatten correlation matrix
}
\examples{
R <- Hmisc::rcorr(as.matrix(mtcars[,1:7])) # sample correlation matrix
flat_corr <- flat_cor_mat(R)
}
