% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get the DHS data used for modelling the proportion of modern contraceptives supplied by the public and private sectors.}
\usage{
get_data(fp2030 = TRUE)
}
\arguments{
\item{fp2030=TRUE}{Filter raw data to only include the Family Planning 2030 focus countries discussed in the Comiskey et al. paper.}
}
\value{
returns the DHS data set used for inputs into the model
}
\description{
Get the DHS data used for modelling the proportion of modern contraceptives supplied by the public and private sectors.
}
