% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_parameters.R
\name{get_local_parameters}
\alias{get_local_parameters}
\title{Get the country-specific median and precision terms for the alpha intercept in local model runs.}
\usage{
get_local_parameters(mycountry, fp2030 = TRUE)
}
\arguments{
\item{mycountry}{The name of country of interest. Default is NULL. For the names of potential countries, review vigentte.}

\item{fp2030}{Default to be TRUE. The country of interest is named as one of the Family Planning 2030 focus countries discussed in the Comiskey et al. paper.}
}
\value{
A list of local parameters to be used to inform the intercept parameter alpha and the global variance-covariance matrix used in the wishart prior of the local_model_run.txt file.
}
\description{
Get the country-specific median and precision terms for the alpha intercept in local model runs.
}
