% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_national_JAGSinput_list.R
\name{get_national_JAGSinput_list}
\alias{get_national_JAGSinput_list}
\title{Combines the data sources to create one JAGS input list}
\usage{
get_national_JAGSinput_list(pkg_data, local = FALSE, mycountry = NULL)
}
\arguments{
\item{pkg_data}{The data list from the 'mcmsupply::get_national_modelinputs' function.}

\item{local}{TRUE/FALSE. Default is FALSE for global runs. Decides if this is a single-country or global run.}

\item{mycountry}{The name of country of interest. Default is NULL. For the names of potential countries, review vigentte.}
}
\value{
returns a list ready for input into the JAGS model
}
\description{
Combines the data sources to create one JAGS input list
}
