% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_national_P_median_quantiles.R
\name{get_national_P_median_quantiles}
\alias{get_national_P_median_quantiles}
\title{Get median and 95\% credible interval for posterior samples of P from national JAGS model}
\usage{
get_national_P_median_quantiles(
  country_index_table,
  method_index_table,
  sector_index_table,
  year_index_table,
  my_model,
  local = FALSE
)
}
\arguments{
\item{country_index_table}{Dataframe with country indexing applied. Used to match estimates to data.}

\item{method_index_table}{Dataframe with method indexing applied. Used to match estimates to data.}

\item{sector_index_table}{Dataframe with sector indexing applied. Used to match estimates to data.}

\item{year_index_table}{Dataframe with time indexing applied. Used to match estimates to data.}

\item{my_model}{JAGS model}
}
\value{
Dataframe of labelled posterior samples with median and 95\% credible intervals estimates.
}
\description{
Get median and 95\% credible interval for posterior samples of P from national JAGS model
}
