% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_point_estimates.R
\name{get_point_estimates}
\alias{get_point_estimates}
\title{Wrapper function to run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model for the national and subnational administration levels}
\usage{
get_point_estimates(main_path, jagsdata, ...)
}
\arguments{
\item{main_path}{Default is "results/". String to indicate where to save results.}

\item{jagsdata}{Output of the \code{mcmsupply::get_modelinputs()} function.}

\item{national}{TRUE/FALSE. Default is TRUE for national administration level data. FALSE retrieves subnational level data.}

\item{local}{TRUE/FALSE. Default is FALSE. local=FALSE retrieves the data for all subnational provinces across all countries. local=TRUE retrieves data for only one country.}

\item{mycountry}{The country name of interest in a local run. You must have local=TRUE for this functionality. A list of possible countries available found in data/mycountries.rda.}
}
\value{
returns the point estimates (median, 80\% and 95\% credible intervals) from the JAGS output
importFrom("stats", "cor", "filter", "lag")
}
\description{
Wrapper function to run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model for the national and subnational administration levels
}
\examples{
National single-country example:
get_point_estimates(national=TRUE, local=TRUE, mycountry="Nepal", main_path = "results/test", jagsdata)

National multi-country example:
get_point_estimates(national=TRUE, local=FALSE, mycountry=NULL, main_path = "results/test", jagsdata)

Subnational single-country example:
get_point_estimates(national=FALSE, local=TRUE, mycountry="Nepal", main_path = "results/test", jagsdata)

Subnational multi-country example:
get_point_estimates(national=FALSE, local=FALSE, mycountry=NULL, main_path = "results/test", jagsdata)

}
