% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subnational_global_P_CM.R
\name{get_subnational_global_P_CM}
\alias{get_subnational_global_P_CM}
\title{Get posterior samples of commercial medical sector from r and z variables of JAGS model}
\usage{
get_subnational_global_P_CM(main_path)
}
\arguments{
\item{main_path}{String. Path where you have set your model results to be saved to.}
}
\value{
Saved samples for commercial medical supply shares.
}
\description{
Get posterior samples of commercial medical sector from r and z variables of JAGS model
}
