% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subnational_global_P_other.R
\name{get_subnational_global_P_other}
\alias{get_subnational_global_P_other}
\title{Get posterior samples of other sector from r and z variables of JAGS model}
\usage{
get_subnational_global_P_other(main_path)
}
\arguments{
\item{main_path}{String. Path where you have set your model results to be saved to.}
}
\value{
Saved samples for other supply shares.
}
\description{
Get posterior samples of other sector from r and z variables of JAGS model
}
