% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_P_public.R
\name{get_subnational_global_P_public}
\alias{get_subnational_global_P_public}
\title{Get posterior samples of public sector from r and z variables of JAGS model}
\usage{
get_subnational_global_P_public(main_path)
}
\arguments{
\item{main_path}{String. Path where you have set your model results to be saved to.}
}
\value{
Saved samples for public supply shares.
}
\description{
Get posterior samples of public sector from r and z variables of JAGS model
}
