% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subnational_local_geodata.R
\name{get_subnational_local_geodata}
\alias{get_subnational_local_geodata}
\title{Get neighbourhood adjacency matrix for your country of interest. Calculate your neighbour count.}
\usage{
get_subnational_local_geodata(mycountry)
}
\arguments{
\item{mycountry}{The country name of interest in a local run. You must have local=TRUE for this functionality. A list of possible countries available found in data/mycountries.rda.}
}
\value{
A list of two matrices. D = Neighbour count matrix, neighbours = Neighbourhood adjacency matrix.
}
\description{
Get neighbourhood adjacency matrix for your country of interest. Calculate your neighbour count.
}
\examples{
get_subnational_local_geodata("Nepal")
}
