% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subnational_local_P_estimates.R
\name{get_subnational_local_P_estimates}
\alias{get_subnational_local_P_estimates}
\title{Get median, 95% and 80% credible intervals for posterior samples of P from local JAGS model}
\usage{
get_subnational_local_P_estimates(
  Psamps,
  param_names,
  subnat_index_table,
  method_index_table,
  sector_index_table,
  year_index_table
)
}
\arguments{
\item{Psamps}{posterior samples of P for one sector from JAGS model}

\item{param_names}{names of the parameters you wish to summarise}

\item{subnat_index_table}{Dataframe with subnational district indexing applied. Used to match estimates to data.}

\item{method_index_table}{Dataframe with method indexing applied. Used to match estimates to data.}

\item{year_index_table}{Dataframe with time indexing applied. Used to match estimates to data.}

\item{sector_type}{String. Name of sector you are interested in. One of either ("Public", "Commercial medical", "Other")}
}
\value{
Dataframe of labelled posterior samples with median, 95% and 80% credible intervals estimates.
}
\description{
Get median, 95% and 80% credible intervals for posterior samples of P from local JAGS model
}
