% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subnational_P_point_estimates.R
\name{get_subnational_P_point_estimates}
\alias{get_subnational_P_point_estimates}
\title{Get the r and z ratio point estimates from the separate chains of the JAGS model runs
R and Z are the intermediate parameters that are used to estimates the final proportions. See the model file for context.}
\usage{
get_subnational_P_point_estimates(
  main_path,
  pkg_data,
  local = FALSE,
  spatial = FALSE,
  mycountry = NULL
)
}
\arguments{
\item{main_path}{String. Path where you have set your model results to be saved to.}

\item{pkg_data}{Output of the `mcmsupplylocal::get_subnational_modelinputs()` function.}

\item{local}{TRUE/FALSE. Default is FALSE. local=FALSE retrieves the data for all subnational provinces across all countries. local=TRUE retrieves data for only one country.}

\item{spatial}{TRUE/FALSE. Default is FALSE. spatial=FALSE retrieves the data for all subnational provinces across all countries without GPS information. spatial=TRUE retrieves for data for countries with GPS information as well as FP source data.}

\item{mycountry}{The country name of interest in a local run. You must have local=TRUE for this functionality. A list of possible countries available found in data/mycountries.rda.}
}
\value{
returns the point estimates for the jags model object
}
\description{
Get the r and z ratio point estimates from the separate chains of the JAGS model runs
R and Z are the intermediate parameters that are used to estimates the final proportions. See the model file for context.
}
