% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_freshness.R
\name{harvest_dhs}
\alias{harvest_dhs}
\title{Harvest DHS Dataset Availability}
\usage{
harvest_dhs(
  url = "https://dhsprogram.com/data/available-datasets.cfm",
  survey_column = "Survey",
  standard_dhs = TRUE
)
}
\arguments{
\item{url}{Character string. URL of the DHS available datasets page.
Default is "https://dhsprogram.com/data/available-datasets.cfm".}

\item{survey_column}{Character string. Name of the column containing
survey information. Default is 'Survey'.}

\item{standard_dhs}{Logical. If TRUE, filters results to include only
Standard DHS surveys. Default is TRUE.}
}
\value{
A tibble containing all tables from the DHS program URL, with
added Country and Year columns extracted from the survey column.
}
\description{
Scrapes the DHS Program website to retrieve information about the
newest available datasets.
}
\examples{
\dontrun{
dhs_data <- harvest_dhs()
}

}
