% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_alpha_region_intercepts.R
\docType{data}
\name{median_alpha_region_intercepts}
\alias{median_alpha_region_intercepts}
\title{The median estimates of the national-level sub-continental, sector-, method-specific intercepts in the multi-country national model.
This array is used to inform the Normal prior of the country-level intercept in the single-country national model.}
\format{
A array of 5 matrices with 2 rows and 5 columns
}
\usage{
median_alpha_region_intercepts
}
\description{
The median estimates of the national-level sub-continental, sector-, method-specific intercepts in the multi-country national model.
This array is used to inform the Normal prior of the country-level intercept in the single-country national model.
}
\keyword{median_alpha_region_intercepts}
