% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_alphacms.R
\docType{data}
\name{median_alphacms}
\alias{median_alphacms}
\title{The median estimates of the subnational-level country, sector-, method-specific intercepts in the multi-country subnational model.
This array is used to inform the Normal prior of the subnational-level intercept in the single-country subnational model.}
\format{
A array of 25 matrices with 2 rows and 5 columns
}
\usage{
median_alphacms
}
\description{
The median estimates of the subnational-level country, sector-, method-specific intercepts in the multi-country subnational model.
This array is used to inform the Normal prior of the subnational-level intercept in the single-country subnational model.
}
\keyword{median_alphacms}
