% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_index.R
\name{method_index_fun}
\alias{method_index_fun}
\title{Indexing Functions for methods used in data.}
\usage{
method_index_fun(my_data, my_methods)
}
\arguments{
\item{my_data}{The data with a  method column you want to index.}

\item{my_methods}{A vector of methods used in the data}
}
\value{
returns data with methods indexed
}
\description{
Indexing Functions for methods used in data.
}
