% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstconnect.R
\name{mstconnect}
\alias{mstconnect}
\title{Connect islands for adjancy matrix GIS
Source: https://gis.stackexchange.com/questions/413159/how-to-assign-a-neighbour-status-to-unlinked-polygons}
\usage{
mstconnect(polys, nb, distance = "centroid")
}
\arguments{
\item{polys}{Spatial polygons to connect}

\item{nb}{Disjointed neighbourhood matrix}

\item{distance}{Default is "centroid".}
}
\description{
Connect islands for adjancy matrix GIS
Source: https://gis.stackexchange.com/questions/413159/how-to-assign-a-neighbour-status-to-unlinked-polygons
}
