% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimated_national_correlations_logitnormal.R
\docType{data}
\name{national_estimated_correlations_logitnormal}
\alias{national_estimated_correlations_logitnormal}
\title{The estimated national-level correlations between the rates of change in methods}
\format{
A array of 2 matrices with 5 rows and 5 columns:
\describe{
\item{row}{Contraceptive method the correlation estimate refers to}
\item{column}{Contraceptive method the correlation estimate refers to}
\item{public_cor}{The estimated correlation between the rates of change in methods supplied by the public sector}
\item{private_cor}{The estimated correlation between the rates of change in methods supplied by the private sector}
}
}
\usage{
national_estimated_correlations_logitnormal
}
\description{
The estimated national-level correlations between the rates of change in methods
}
\keyword{national_estimated_correlations_logitnormal}
