% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_FPsource_data.R
\docType{data}
\name{national_FPsource_data}
\alias{national_FPsource_data}
\title{DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the national level}
\format{
\subsection{\code{national_FPsource_data}}{

A data frame with 2448 rows and 9 columns:
\describe{
\item{year}{Year of survey}
\item{Country}{Country names}
\item{Super_region}{Sub-continent names}
\item{Method}{Contraceptive method name}
\item{average_year}{Average year of the survey}
\item{sector_category}{Name of sector}
\item{proportion}{Proportion supplied by the sector}
\item{SE.proportion}{Standard error associated with the proportion}
\item{n}{Sample size associated with the observation}
}
}
}
\source{
On request from DHS microdata - using the womens individuals recode file. Contact details found at https://dhsprogram.com/data/dataset_admin/login_main.cfm
}
\usage{
national_FPsource_data
}
\description{
DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the national level
}
\keyword{datasets}
