% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_FPsource_format.R
\docType{data}
\name{national_FPsource_format}
\alias{national_FPsource_format}
\title{A checklist for ensuring national-level custom data is appropriate to be used for estimation}
\format{
A list of requirements for custom data to meet before being accepted into the model
\describe{
\item{Country}{Country name variable classification}
\item{Method}{Method name variable classification}
\item{average_year}{average_year variable classification}
\item{sector_category}{sector_category name variable classification}
\item{proportion}{proportion variable classification}
\item{SE.proportion}{SE.proportion variable classification}
\item{n}{n variable classification}
}
}
\usage{
national_FPsource_format
}
\description{
A checklist for ensuring national-level custom data is appropriate to be used for estimation
}
\keyword{national_FPsource_format}
