% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_FPsource_VARCOV_bivarlogitnormal.R
\docType{data}
\name{national_FPsource_VARCOV_bivarlogitnormal}
\alias{national_FPsource_VARCOV_bivarlogitnormal}
\title{An array of variance-covariance matrices transformed onto the logit-scale via the delta method. Each matrix corresponds to the DHS survey logit-transformed observations for the proportion of modern contraceptives supplied by the public and private sectors at the national level}
\format{
\subsection{\code{national_FPsource_VARCOV_bivarlogitnormal}}{

An array of 2x2 matrices for each of the 608 observations in the national_FPsource_data.
Each 2x2 array corresponds to the variance of the public and private sectors on the diagonal and their corresponding covariances on the off-diagonal.
}
}
\source{
The variance-covariance matrices are calculated using the survey R package:
prop_mat <- svyby(~I(sector_categories), ~I(modern_method_source), design=d.s, svymean, covmat=TRUE)
vcov_matrix <- vcov(prop_mat) function.
}
\usage{
national_FPsource_VARCOV_bivarlogitnormal
}
\description{
An array of variance-covariance matrices transformed onto the logit-scale via the delta method. Each matrix corresponds to the DHS survey logit-transformed observations for the proportion of modern contraceptives supplied by the public and private sectors at the national level
}
\keyword{datasets}
