% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_inv_sigma_delta_hat_bivarlogitnormal.R
\docType{data}
\name{national_inv_sigma_delta_hat_bivarlogitnorm}
\alias{national_inv_sigma_delta_hat_bivarlogitnorm}
\title{The median estimate for the national-level variance-covariance matrix of the delta.k terms in the multi-country national model.}
\format{
A array of 2 matrices with 5 rows and 5 columns:
\describe{
\item{Array 1}{Estimated public sector variance-covaraince matrix}
\item{Array 2}{Estimated private sector variance-covaraince matrix}
}
}
\usage{
national_inv_sigma_delta_hat_bivarlogitnorm
}
\description{
The median estimate for the national-level variance-covariance matrix of the delta.k terms in the multi-country national model.
}
\keyword{national_inv_sigma_delta_hat_bivarlogitnorm}
