% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_tau_alpha_cms_hat_logitnormal.R
\docType{data}
\name{national_tau_alpha_cms_hat_logitnormal}
\alias{national_tau_alpha_cms_hat_logitnormal}
\title{The median estimates of the precision for the national-level country, sector-, method-specific intercepts in the multi-country national model.
This vector is used to inform the precision in the Normal prior of the national-level intercept in single-country national models.}
\format{
A vector of two precision estimates
}
\usage{
national_tau_alpha_cms_hat_logitnormal
}
\description{
The median estimates of the precision for the national-level country, sector-, method-specific intercepts in the multi-country national model.
This vector is used to inform the precision in the Normal prior of the national-level intercept in single-country national models.
}
\keyword{national_tau_alpha_cms_hat_logitnormal}
