% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_theta_rms_hat_bivarlogitnormal.R
\docType{data}
\name{national_theta_rms_hat_bivarlogitnorm}
\alias{national_theta_rms_hat_bivarlogitnorm}
\title{The median estimates of the national-level sub-continental, sector-, method-specific intercepts in the multi-country national model.
This array is used to inform the Normal prior of the country-level intercept in the single-country national model.}
\format{
A array of 5 matrices with 2 rows and 5 columns
}
\usage{
national_theta_rms_hat_bivarlogitnorm
}
\description{
The median estimates of the national-level sub-continental, sector-, method-specific intercepts in the multi-country national model.
This array is used to inform the Normal prior of the country-level intercept in the single-country national model.
}
\keyword{national_theta_rms_hat_bivarlogitnorm}
