% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_varcov_order_bivarlogitnormal.R
\docType{data}
\name{national_varcov_order_bivarlogitnormal}
\alias{national_varcov_order_bivarlogitnormal}
\title{The order of observations to join the variance-covariance array data and the DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the national level}
\format{
\subsection{\code{national_varcov_order_bivarlogitnormal}}{

A data frame with 558 rows and 3 columns:
\describe{
\item{Country}{Country names}
\item{average_year}{Average year of the survey}
\item{Method}{Contraceptive method name}
}
}
}
\source{
On request from DHS microdata - using the womens individuals recode file. Contact details found at https://dhsprogram.com/data/dataset_admin/login_main.cfm
}
\usage{
national_varcov_order_bivarlogitnormal
}
\description{
The order of observations to join the variance-covariance array data and the DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the national level
}
\keyword{datasets}
