% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Wrapper function to plot the JAGS estimates}
\usage{
plot_estimates(main_path, vis_path, jagsdata, ...)
}
\arguments{
\item{main_path}{String. Path where you have set your model results to be saved to.}

\item{vis_path}{String. Path where your visualisations will be saved to.}

\item{jagsdata}{Output of the \code{mcmsupply::get_modelinputs()} function.}

\item{national}{TRUE/FALSE. Default is TRUE for national administration level data. FALSE retrieves subnational level data.}

\item{local}{TRUE/FALSE. Default is FALSE. local=FALSE retrieves the data for all subnational provinces across all countries. local=TRUE retrieves data for only one country.}

\item{mycountry}{The country name of interest in a local run. You must have local=TRUE for this functionality. A list of possible countries available found in data/mycountries.rda.}
}
\description{
Wrapper function to plot the JAGS estimates
}
\examples{
National single-country example:
plot_estimates(main_path = "results/test", vis_path  = "visualisations/test", jagsdata, ...)

National multi-country example:
plot_estimates(main_path = "results/test", vis_path  = "visualisations/test", jagsdata, ...)

Subnational single-country example:
plot_estimates(main_path = "results/test", vis_path  = "visualisations/test", jagsdata, ...)

Subnational multi-country example:
plot_estimates(main_path = "results/test", vis_path  = "visualisations/test", jagsdata, ...)

}
