% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_national_point_estimates.R
\name{plot_national_point_estimates}
\alias{plot_national_point_estimates}
\title{Plot median and 95\% credible intervals for posterior samples of P from JAGS model with the relevant survey data}
\usage{
plot_national_point_estimates(
  main_path,
  vis_path,
  pkg_data,
  local = FALSE,
  mycountry = NULL
)
}
\arguments{
\item{main_path}{String. Path where you have set your model results to be saved to.}

\item{vis_path}{String. Path where your visualisations will be saved to.}

\item{pkg_data}{Output of the \code{mcmsupplylocal::get_subnational_modelinputs()} function.}

\item{local}{TRUE/FALSE. Default is FALSE. local=FALSE retrieves the data for all subnational provinces across all countries. local=TRUE retrieves data for only one country.}

\item{mycountry}{The country name of interest in a local run. You must have local=TRUE for this functionality. A list of possible countries available found in data/mycountries.rda.}
}
\value{
Data frame of labelled posterior samples with median, 95\% and 80\% credible intervals estimates.
}
\description{
Plot median and 95\% credible intervals for posterior samples of P from JAGS model with the relevant survey data
}
