% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point_estimates.R
\name{plot_point_estimates}
\alias{plot_point_estimates}
\title{Wrapper function to plot the JAGS estimates}
\usage{
plot_point_estimates(
  national = TRUE,
  main_path,
  vis_path,
  pkg_data,
  local = FALSE,
  mycountry = NULL
)
}
\arguments{
\item{national}{TRUE/FALSE. Default is TRUE for national administration level data. FALSE retrieves subnational level data.}

\item{main_path}{String. Path where you have set your model results to be saved to.}

\item{vis_path}{String. Path where your visualisations will be saved to.}

\item{pkg_data}{Output of the \code{mcmsupply::get_subnational_modelinputs()} function.}

\item{local}{TRUE/FALSE. Default is FALSE. local=FALSE retrieves the data for all subnational provinces across all countries. local=TRUE retrieves data for only one country.}

\item{mycountry}{The country name of interest in a local run. You must have local=TRUE for this functionality. A list of possible countries available found in data/mycountries.rda.}
}
\description{
Wrapper function to plot the JAGS estimates
}
\examples{
National single-country example:
plot_point_estimates(national=TRUE, main_path = "results/test", vis_path  = "visualisations/test", pkg_data, local=TRUE, mycountry="Nepal")

National multi-country example:
plot_point_estimates(national=TRUE,  main_path = "results/test", vis_path  = "visualisations/test", pkg_data, local=FALSE, mycountry=NULL)

Subnational single-country example:
plot_point_estimates(national=FALSE, main_path = "results/test", vis_path  = "visualisations/test", pkg_data, local=TRUE, mycountry="Nepal")

Subnational multi-country example:
plot_point_estimates(national=FALSE, main_path = "results/test", vis_path  = "visualisations/test", pkg_data, local=FALSE, mycountry=NULL)

}
