% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_alpha_country_intercepts.R
\docType{data}
\name{precision_alpha_country_intercepts}
\alias{precision_alpha_country_intercepts}
\title{The median estimates of the precision for the national-level country, sector-, method-specific intercepts in the multi-country national model.
This vector is used to inform the precision in the Normal prior of the national-level intercept in single-country national models.}
\format{
A vector of two precision estimates
}
\usage{
precision_alpha_country_intercepts
}
\description{
The median estimates of the precision for the national-level country, sector-, method-specific intercepts in the multi-country national model.
This vector is used to inform the precision in the Normal prior of the national-level intercept in single-country national models.
}
\keyword{precision_alpha_country_intercepts}
