% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_index.R
\name{region_index_fun}
\alias{region_index_fun}
\title{Indexing function for regions used in data.}
\usage{
region_index_fun(my_data, n_region)
}
\arguments{
\item{my_data}{The data with a Region column you want to index.}

\item{n_region}{A vector of regions used in the data}
}
\value{
returns data with region indexed
}
\description{
Indexing function for regions used in data.
}
