% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags_model.R
\name{run_jags_model}
\alias{run_jags_model}
\title{Run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model}
\usage{
run_jags_model(jagsdata, jagsparams)
}
\arguments{
\item{mydata}{The DHS data SE_source_data_20.RDS from /data folder. Contains data for countries participating in the FP2030 initiative.}

\item{nseg}{Ghe number of knots to use in the splines. Default is 12.#'}
}
\value{
returns the jags model object
importFrom("stats", "cor", "filter", "lag")
}
\description{
Run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model
}
