% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_national_jags_model.R
\name{run_national_jags_model}
\alias{run_national_jags_model}
\title{Run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model}
\usage{
run_national_jags_model(jagsdata, jagsparams, local_run = FALSE, mycountry)
}
\arguments{
\item{jagsdata}{The inputs for the JAGS model}

\item{jagsparams}{The parameters of the JAGS model you wish to review}

\item{local_run}{TRUE/FALSE. Default is FALSE for global runs. Decides if this is a single-country or global run.}

\item{mycountry}{The name of the country you are estimating for.}
}
\value{
returns the jags model object
importFrom("stats", "cor", "filter", "lag")
}
\description{
Run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model
}
