\name{SE_source_data_20}
\docType{data}
\alias{SE_source_data_20}
\title{Family Planning Source Data}
\description{
  This data set provides the proportion of modern contraceptive methods supplied
  by the public and private sectors over time for a selection of countries with DHS surveys.
}
\usage{SE_source_data_20}
\format{A datatable containing 8 variables and 1324 observations}
\source{Demographic and Health Survey Data}
\references{
  ICF. STATcompiler. The DHS Program STATcompiler. Funded by USAID. (2012).
}
\keyword{datasets}
