% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_index.R
\name{sector_index_fun}
\alias{sector_index_fun}
\title{Indexing function for sectors used in data.}
\usage{
sector_index_fun(my_data, my_sectors)
}
\arguments{
\item{my_data}{The data with a sector_category column you want to index.}

\item{my_sectors}{A vector of sectors used in the data}
}
\value{
returns data with sectors indexed
}
\description{
Indexing function for sectors used in data.
}
