% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_jags_data.R
\name{set_up_jags_data}
\alias{set_up_jags_data}
\title{Clean the raw input DHS data so it is suitable to be used in the jags model.
Primarily this function addresses missing observations and standard errors (SE), replaces extremely tiny SE with reasonable SE}
\usage{
set_up_jags_data(mydata)
}
\arguments{
\item{mydata}{The DHS data SE_source_data_20.RDS from /data folder. Contains data for countries participating in the FP2030 initiative.}
}
\value{
returns a cleaned DHS dataset
}
\description{
Clean the raw input DHS data so it is suitable to be used in the jags model.
Primarily this function addresses missing observations and standard errors (SE), replaces extremely tiny SE with reasonable SE
}
