% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_method_names.R
\name{standard_method_names}
\alias{standard_method_names}
\title{Apply standard naming for modern contraceptive methods to dataset}
\usage{
standard_method_names(my_data)
}
\arguments{
\item{my_data}{A dataframe or tibble containing a column called "Method" which contains the names of modern contraceptive methods}
}
\value{
A dataframe or tibble that has standardised naming applied to the Method column
}
\description{
Apply standard naming for modern contraceptive methods to dataset
}
