% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnat_FPsource_data.R
\docType{data}
\name{subnat_FPsource_data}
\alias{subnat_FPsource_data}
\title{DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the subnational adminstration level}
\format{
\subsection{\code{subnat_FPsource_data}}{

A data frame with 13504 rows and 8 columns:
\describe{
\item{Country}{Country names}
\item{Region}{Subnational region names}
\item{Method}{Contraceptive method name}
\item{average_year}{Average year of the survey}
\item{sector_categories}{Name of sector}
\item{proportion}{Proportion supplied by the sector}
\item{SE.proportion}{Standard error associated with the proportion}
\item{n}{Sample size associated with the observation}
}
}
}
\source{
On request from DHS microdata - using the womens individuals recode file. Contact details found at https://dhsprogram.com/data/dataset_admin/login_main.cfm
}
\usage{
subnat_FPsource_data
}
\description{
DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the subnational adminstration level
}
\keyword{datasets}
