% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnat_index_fun.R
\name{subnat_index_fun}
\alias{subnat_index_fun}
\title{Apply subnational indexing to data}
\usage{
subnat_index_fun(my_data, my_subnat, my_country)
}
\arguments{
\item{my_data}{The sub-national family planning source data for the country of interest.}

\item{my_subnat}{A vector of the provinces/region names in your country of interest.}

\item{my_country}{The name of your country of interest. Must be a vector the same length as my_subnat.}
}
\value{
Dataframe with subnational indexing applied
}
\description{
Apply subnational indexing to data
}
\examples{
country_subnat_tbl <- mydata \%>\%
dplyr::group_by(Country, Region) \%>\%
dplyr::select(Country, Region) \%>\%
dplyr::distinct() # table of country and regions (removes repeats in region names)

subnat_index_fun(mydata, country_subnat_tbl$Region, country_subnat_tbl$Country)
}
