% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnational_alpha_cms_hat.R
\docType{data}
\name{subnational_alpha_cms_hat}
\alias{subnational_alpha_cms_hat}
\title{The median estimates of the subnational-level country, sector-, method-specific intercepts in the multi-country subnational model.
This array is used to inform the Normal prior of the subnational-level intercept in the single-country subnational model.}
\format{
A array of 23 matrices with 2 rows and 5 columns
}
\usage{
subnational_alpha_cms_hat
}
\description{
The median estimates of the subnational-level country, sector-, method-specific intercepts in the multi-country subnational model.
This array is used to inform the Normal prior of the subnational-level intercept in the single-country subnational model.
}
\keyword{subnational_alpha_cms_hat}
