% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnational_tau_alpha_pms_hat.R
\docType{data}
\name{subnational_tau_alpha_pms_hat}
\alias{subnational_tau_alpha_pms_hat}
\title{subnational_tau_alpha_pms_hat
The median estimates of the precision for the subnational-level country, sector-, method-specific intercepts in the multi-country subnational model.
This vector is used to inform the precision in the Normal prior of the subnational-level intercept in single-country subnational models.}
\format{
A vector of two precision estimates
}
\usage{
subnational_tau_alpha_pms_hat
}
\description{
subnational_tau_alpha_pms_hat
The median estimates of the precision for the subnational-level country, sector-, method-specific intercepts in the multi-country subnational model.
This vector is used to inform the precision in the Normal prior of the subnational-level intercept in single-country subnational models.
}
\keyword{subnational_tau_alpha_pms_hat}
