% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superregion_index_fun.R
\name{superregion_index_fun}
\alias{superregion_index_fun}
\title{Apply superregion indexing to data}
\usage{
superregion_index_fun(my_data, n_region)
}
\arguments{
\item{my_data}{The sub-national family planning source data for the country of interest.}

\item{n_region}{A vector of the super-regions listed in the dataset}
}
\value{
Dataframe with super-region indexing applied
}
\description{
Apply superregion indexing to data
}
\examples{
n_superreg <- unique(mydata$Super_region)
mydata <- superregion_index_fun(mydata, n_superreg)
}
