% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_alpha_pms_hat.R
\docType{data}
\name{tau_alphapms_subnationalmod}
\alias{tau_alphapms_subnationalmod}
\title{The median estimates of the precision for the subnational-level country, sector-, method-specific intercepts in the multi-country subnational model.
This vector is used to inform the precision in the Normal prior of the subnational-level intercept in single-country subnational models.}
\format{
A vector of two precision estimates
}
\usage{
tau_alphapms_subnationalmod
}
\description{
The median estimates of the precision for the subnational-level country, sector-, method-specific intercepts in the multi-country subnational model.
This vector is used to inform the precision in the Normal prior of the subnational-level intercept in single-country subnational models.
}
\keyword{tau_alphapms_subnationalmod}
