% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_national_FPsource_data.R
\docType{data}
\name{trunc_national_FPsource_data}
\alias{trunc_national_FPsource_data}
\title{A small sample dataset of four countries for illustration purposes.
Countries included are Benin, Sierra Leone, Madagascar, and Zimbabwe.
DHS survey observations for the proportion of modern contraceptives supplied
by the public and private sectors at the national level}
\format{
\subsection{\code{trunc_national_FPsource_data}}{

A data frame with 179 rows and 11 columns:
\describe{
\item{year}{Year of survey}
\item{Country}{Country names}
\item{Method}{Contraceptive method name}
\item{average_year}{Average year of the survey}
\item{sector_category}{Name of sector}
\item{proportion}{Proportion supplied by the sector}
\item{SE.proportion}{Standard error associated with the proportion}
\item{country_code}{Letter code for country name}
\item{n}{Sample size associated with the observation}
\item{check_sum}{Evaluating the total of the proportions across
the three sectors for a given year, method and country}
\item{Method_collapse}{Standardised method names}
}
}
}
\source{
On request from DHS surveys, using the womens individual recode file
}
\usage{
trunc_national_FPsource_data
}
\description{
Contact details found at:
https://dhsprogram.com/data/dataset_admin/login_main.cfm
}
\keyword{datasets}
