% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_bbase_precise.R
\name{bs_bbase_precise}
\alias{bs_bbase_precise}
\title{Get precisely aligned basis functions}
\usage{
bs_bbase_precise(
  x = x,
  lastobs = max(x),
  xl = min(x),
  xr = max(x),
  nseg = nseg,
  deg = 3
)
}
\arguments{
\item{x}{The vector of years you wish to create your basis functions over}

\item{lastobs}{The year of the most recent survey you wish to align the knots with. Default is max(x).}

\item{xl}{Default is xl = min(x)}

\item{xr}{Default is xr = max(x)}

\item{nseg}{Number of knots you wish to use}

\item{deg}{The degree of the polynomial. Default is 3.}
}
\value{
B.ik is a matrix, each row is one observation, each column is one B-spline.
knots.k is a vector of transformed knots.
Kstar is the knot point of last observation
}
\description{
Get precisely aligned basis functions
}
\examples{
all_years <- seq(from = 1990, to = 2030.5, by=0.5)
bs_bbase_precise(all_years, lastobs=2014.5, nseg = 12)
}
