% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_index.R
\name{country_index_fun}
\alias{country_index_fun}
\alias{continent_index_fun}
\title{Indexing function for countries used in data.}
\usage{
country_index_fun(my_data, my_countries)

continent_index_fun(my_data)
}
\arguments{
\item{my_data}{The data with a Country column you want to index.}

\item{my_countries}{A vector of country names used in the data}
}
\value{
returns table with countries indexed
}
\description{
Indexing function for countries used in data.
}
\section{Functions}{
\itemize{
\item \code{country_index_fun()}: Calculates the country index for named countries

\item \code{continent_index_fun()}: Calculates the continent index for named continents

}}
