% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_index_fun.R
\name{country_index_fun}
\alias{country_index_fun}
\title{Apply country indexing to data}
\usage{
country_index_fun(my_data, my_country)
}
\arguments{
\item{my_data}{The sub-national family planning source data for the country of interest.}

\item{my_country}{The name of your country of interest.}
}
\value{
Dataframe with country indexing applied
}
\description{
Apply country indexing to data
}
\examples{
country_index_fun(mydata, "Nepal")
}
