% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_national_data.R
\name{get_national_data}
\alias{get_national_data}
\title{Get the DHS data used for modelling the proportion of modern contraceptives supplied by the public and private sectors at the national level.}
\usage{
get_national_data(local = FALSE, mycountry = NULL, fp2030 = TRUE)
}
\arguments{
\item{local}{TRUE/FALSE. Default is FALSE for global runs. Decides if this is a single-country or global run.}

\item{mycountry}{The name of country of interest. Default is NULL. For the names of potential countries, review vigentte.}

\item{fp2030=TRUE}{Filter raw data to only include the Family Planning 2030 focus countries discussed in the Comiskey et al. paper.}
}
\value{
returns the DHS data set used for inputs into the model
}
\description{
Get the DHS data used for modelling the proportion of modern contraceptives supplied by the public and private sectors at the national level.
}
