% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_point_estimates.R
\name{get_point_estimates}
\alias{get_point_estimates}
\title{Wrapper function to run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model for the national and subnational administration levels}
\usage{
get_point_estimates(main_path = NULL, jagsdata, ...)
}
\arguments{
\item{main_path}{Default is "results/". String to indicate where to save results.}

\item{jagsdata}{Output of the \code{mcmsupply::get_modelinputs()} function.}

\item{...}{Arguments from the \code{mcmsupply::get_modelinputs()} function.}
}
\value{
returns the point estimates (median, 80\% and 95\% credible intervals) from the JAGS output
importFrom("stats", "cor", "filter", "lag")
}
\description{
Wrapper function to run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model for the national and subnational administration levels
}
\examples{
National single-country example:
get_point_estimates(main_path = "results/test", jagsdata)

National multi-country example:
get_point_estimates(main_path = "results/test", jagsdata)

Subnational single-country example:
get_point_estimates(main_path = "results/test", jagsdata)

Subnational multi-country example:
get_point_estimates(main_path = "results/test", jagsdata)

}
