% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_index_fun.R
\name{method_index_fun}
\alias{method_index_fun}
\title{Apply method indexing to data}
\usage{
method_index_fun(my_data, my_methods)
}
\arguments{
\item{my_data}{The sub-national family planning source data for the country of interest.}

\item{my_methods}{A vector of your contraceptive methods in the order you wished them indexed.}
}
\value{
Dataframe with method indexing applied
}
\description{
Apply method indexing to data
}
\examples{
  n_method <- c("Female Sterilization", "Implants", "Injectables", "IUD", "OC Pills" ) # As per the method correlation matrix
method_index_fun(mydata, n_method)
}
