% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_FPsource_data.R
\docType{data}
\name{national_FPsource_data}
\alias{national_FPsource_data}
\title{DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the national level}
\format{
\subsection{\code{national_FPsource_data}}{

A data frame with 491 rows and 15 columns:
\describe{
\item{Country}{Country names}
\item{Method}{Contraceptive method name}
\item{average_year}{Average year of the survey}
\item{year}{Year of the survey}
\item{country_code}{ISO country codes}
\item{Public}{Proportion supplied by the Public sector}
\item{se.Public}{Standard error of proportion supplied by the Public sector}
\item{Public_n}{Sample size used to calculate proportion supplied by the Public sector}
\item{Commercial_medical}{Proportion supplied by the private Commercial medical sector}
\item{se.Commercial_medical}{Standard error of proportion supplied by the private Commercial medical sector}
\item{Commercial_medical_n}{Sample size used to calculate proportion supplied by the private Commercial medical sector}
\item{Other}{Proportion supplied by the private Other sector}
\item{se.Other}{Standard error of proportion supplied by the private Other sector}
\item{Other_n}{Sample size used to calculate proportion supplied by the private Other sector}
\item{check_sum}{Check to see if proportions sum to 1}
}
}
}
\source{
On request from DHS microdata - using the womens individuals recode file. Contact details found at https://dhsprogram.com/data/dataset_admin/login_main.cfm
}
\usage{
national_FPsource_data
}
\description{
DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the national level
}
\keyword{datasets}
