% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags_model.R
\name{run_jags_model}
\alias{run_jags_model}
\title{Wrapper function to run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model for the national and subnational administration levels}
\usage{
run_jags_model(
  jagsdata,
  jagsparams = NULL,
  main_path,
  n_iter = 80000,
  n_burnin = 10000,
  n_thin = 35,
  ...
)
}
\arguments{
\item{jagsdata}{The inputs for the JAGS model}

\item{jagsparams}{The parameters of the JAGS model you wish to review}

\item{main_path}{Default is "results/". String to indicate where to save results.}

\item{n_iter}{Default is 80000. Number of itterations to do in JAGS model.}

\item{n_burnin}{Default is 10000. Number of samples to burn-in in JAGS model.}

\item{n_thin}{Default is 35. Number of samples to thin by in JAGS model.}

\item{...}{Argumentsfrom the \code{mcmsupply::get_modelinputs()} function.}
}
\value{
returns the jags model object
importFrom("stats", "cor", "filter", "lag")
}
\description{
Wrapper function to run the jags model for estimating the proportion of modern contraceptive methods supplied by the public & private Sectors using a Bayesian hierarchical penalized spline model for the national and subnational administration levels
}
\examples{
National single-country example:
run_jags_model(jagsdata, jagsparams = NULL, main_path, n_iter = 80000, n_burnin = 10000, n_thin = 35, ...)

National multi-country example:
run_jags_model(jagsdata, jagsparams = NULL, main_path, n_iter = 80000, n_burnin = 10000, n_thin = 35, ...)

Subnational single-country example:
run_jags_model(jagsdata, jagsparams = NULL, main_path, n_iter = 80000, n_burnin = 10000, n_thin = 35, ...)

Subnational multi-country example:
run_jags_model(jagsdata, jagsparams = NULL, main_path, n_iter = 80000, n_burnin = 10000, n_thin = 35, ...)

To change from the default parameters monitored in the JAGS model:
myjagsparams <-c("P","alpha_pms")
run_jags_model(jagsdata, jagsparams = myjagsparams, main_path, n_iter = 80000, n_burnin = 10000, n_thin = 35, ...)
}
