% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_method_names.R
\name{standard_method_names}
\alias{standard_method_names}
\title{Standardise the contraceptive method names}
\usage{
standard_method_names(my_data)
}
\arguments{
\item{my_data}{The input data with a Method column to standardise}
}
\value{
Original input dataframe with Method column standardised.
}
\description{
Standardise the contraceptive method names
}
\examples{
data_standard <- standard_method_names(mydata)
}
