% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnat_FPsource_data.R
\docType{data}
\name{subnat_FPsource_data}
\alias{subnat_FPsource_data}
\title{DHS survey observations for the proportion of modern contraceptives
supplied by the public and private sectors at the subnational
administration level.}
\format{
\subsection{\code{subnat_FPsource_data}}{

A data frame with 6940 rows and 8 columns:
\describe{
\item{Country}{Country names}
\item{Region}{Subnational region names}
\item{Method}{Contraceptive method name}
\item{average_year}{Average year of the survey}
\item{sector_categories}{Name of sector}
\item{proportion}{Proportion supplied by the sector}
\item{SE.proportion}{Standard error associated with the proportion}
\item{n}{Sample size associated with the observation}
}
}
}
\source{
On request from IPUMS - https://www.idhsdata.org/idhs/index.shtml
}
\usage{
subnat_FPsource_data
}
\description{
Elizabeth Heger Boyle, Miriam King and Matthew Sobek.
IPUMS-Demographic and Health Surveys: Version 9 (dataset).
IPUMS and ICF, 2022. https://doi.org/10.18128/D080.V9
}
\keyword{datasets}
